Music: library loading tweaks
