+++ apps/src/music/redux/musicRedux.ts
@@ -24,8 +24,6 @@ export enum InstructionsPosition {
 }
 
 export interface MusicState {
-  /** Current music library name */
-  libraryName: string | null;
   /** Current pack ID, if a specific restricted pack from the current music library is selected */
   packId: string | null;
   /** If the song is currently playing */
@@ -77,7 +75,6 @@ export interface MusicState {
 }
 
 const initialState: MusicState = {
-  libraryName: null,
   packId: null,
   isPlaying: false,
   currentPlayheadPosition: 0,
@@ -113,9 +110,6 @@ const musicSlice = createSlice({
   name: 'music',
   initialState,
   reducers: {
-    setLibraryName: (state, action: PayloadAction<string>) => {
-      state.libraryName = action.payload;
-    },
     setPackId: (state, action: PayloadAction<string>) => {
       state.packId = action.payload;
     },
@@ -301,7 +295,6 @@ export const getCurrentlyPlayingBlockIds = (state: {
 registerReducers({music: musicSlice.reducer});
 
 export const {
-  setLibraryName,
   setPackId,
   setIsPlaying,
   setCurrentPlayheadPosition,
