--- apps/src/music/utils/Loader.ts
@@ -1,47 +1,0 @@
-// Utilities for retrieving various types of data for Music Lab.
-
-import HttpClient from '@cdo/apps/util/HttpClient';
-
-import {getBaseAssetUrl} from '../appConfig';
-import MusicLibrary, {
-  LibraryJson,
-  LibraryValidator,
-} from '../player/MusicLibrary';
-
-const AppConfig = require('../appConfig').default;
-
-// This value can be modifed each time we know that there is an important new version
-// of the library on S3, to help bypass any caching of an older version.
-const requestVersion = 'launch2024-0';
-
-/**
- * Loads a sound library JSON file.
- *
- * @param libraryName specific library to load. If a library is specified by
- * URL param, that will take precedence.
- * @returns the Music Library
- */
-export const loadLibrary = async (
-  libraryName: string
-): Promise<MusicLibrary> => {
-  const libraryParameter = AppConfig.getValue('library') || libraryName;
-  const libraryFilename = `music-library-${libraryParameter}`;
-
-  if (AppConfig.getValue('local-library') === 'true') {
-    const localLibrary = require(`@cdo/static/music/${libraryFilename}.json`);
-    return new MusicLibrary(
-      'local-' + libraryFilename,
-      localLibrary as LibraryJson
-    );
-  } else {
-    const libraryJsonResponse = await HttpClient.fetchJson<LibraryJson>(
-      getBaseAssetUrl() +
-        libraryFilename +
-        '.json' +
-        (requestVersion ? `?version=${requestVersion}` : ''),
-      {},
-      LibraryValidator
-    );
-    return new MusicLibrary(libraryFilename, libraryJsonResponse.value);
-  }
-};
